Sweep the redis-py tests across 5/6/7/8, and run chaos on redis-py 8
The socket-timeout regression tests are client-version-specific and need a
real standalone Redis, so they were skipped on the Memory and cluster legs --
which left those legs short of 100% coverage (the suite measures tests/ too).

Move them into their own tests/redis_py module that the main matrix ignores
(the same way tests/cli is handled, both ignored at collection and omitted
from .coveragerc-core). The dedicated test-redis-py job now runs just that
focused suite, which is cheap enough to sweep every redis-py major we support
(5, 6, 7, 8) against a fixed Redis 8.6 instead of the whole suite on 5 and 8.

Also point the chaos workflow at the latest redis-py (`--upgrade-package
redis`), so the fault-injection and signal runs exercise the socket_timeout
fix under redis-py 8 too.
