+++ .github/workflows/ci.yml
@@ -92,7 +92,7 @@ jobs:
       - name: Run tests
         env:
           REDIS_VERSION: ${{ matrix.backend.redis-version }}
-        run: uv run pytest --ignore=tests/cli --cov-branch --cov-config=.coveragerc-core --cov-fail-under=100 --cov-report=xml --cov-report=term-missing:skip-covered --junitxml=junit.xml -o junit_family=legacy ${{ matrix.pytest-args }}
+        run: uv run pytest --ignore=tests/cli --ignore=tests/redis_py --cov-branch --cov-config=.coveragerc-core --cov-fail-under=100 --cov-report=xml --cov-report=term-missing:skip-covered --junitxml=junit.xml -o junit_family=legacy ${{ matrix.pytest-args }}
 
       - name: Upload coverage to Codecov
         if: ${{ !cancelled() }}
@@ -182,7 +182,7 @@ jobs:
         env:
           REDIS_VERSION: memory
         run: >
-          uv run pytest --ignore=tests/cli
+          uv run pytest --ignore=tests/cli --ignore=tests/redis_py
           --cov-branch --cov-config=.coveragerc-core
           --cov-fail-under=100
           --cov-report=xml --cov-report=term-missing:skip-covered
@@ -209,12 +209,15 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        # The client is what varies here, so we pin one Python and one real
-        # Redis and sweep the redis-py majors we support.  A real backend is
-        # required: the bug this guards (issue #425) is a client socket
-        # timeout, which the memory backend can't exercise.
+        # The client is what varies, so we pin one Python and one real Redis
+        # and sweep the redis-py majors we support.  These run only the
+        # focused tests/redis_py suite (the client-version-sensitive socket
+        # behavior), so all four majors stay cheap.  A real, standalone Redis
+        # is required: the bug this guards (#425) is a client socket timeout.
         redis-py:
           - { name: "5", spec: ">=5,<6" }
+          - { name: "6", spec: ">=6,<7" }
+          - { name: "7", spec: ">=7,<8" }
           - { name: "8", spec: ">=8" }
     steps:
       - uses: actions/checkout@v6
@@ -229,22 +232,14 @@ jobs:
       - name: Install dependencies
         run: uv sync --upgrade-package 'redis${{ matrix.redis-py.spec }}'
 
-      - name: Run tests
+      - name: Run redis-py tests
         env:
           REDIS_VERSION: "8.6"
         run: >
-          uv run pytest --ignore=tests/cli
-          --cov-branch --cov-config=.coveragerc-core
-          --cov-fail-under=100
-          --cov-report=xml --cov-report=term-missing:skip-covered
+          uv run pytest tests/redis_py/
+          --cov=tests/redis_py --cov-branch --cov-fail-under=100
+          --cov-report=term-missing:skip-covered
           --junitxml=junit.xml -o junit_family=legacy
-
-      - name: Upload coverage to Codecov
-        if: ${{ !cancelled() }}
-        uses: codecov/codecov-action@v6
-        with:
-          token: ${{ secrets.CODECOV_TOKEN }}
-          flags: redis-py-${{ matrix.redis-py.name }}
 
       - name: Upload test results to Codecov
         if: ${{ !cancelled() }}
