+++ tests/test_docket_keys.py
@@ -2,14 +2,11 @@
 
 # pyright: reportPrivateUsage=false
 
-import time
-
 import pytest
 import redis.exceptions
 
 from docket._redis import RedisConnection
 from docket.docket import Docket
-from tests.conftest import skip_cluster, skip_memory
 
 
 # Tests for prefix property and key() method
@@ -230,36 +227,3 @@ def test_redis_connection_normalized_url_returns_original_for_non_cluster():
 
     connection2 = RedisConnection("memory://")
     assert connection2._normalized_url() == "memory://"
-
-
-# Tests for blocking-read socket timeout (issue #425)
-
-
-@skip_memory
-@skip_cluster
-async def test_connection_pool_disables_socket_read_timeout(redis_url: str):
-    """Docket's blocking reads must not be cut short by a client read timeout.
-
-    redis-py 8 defaults socket_timeout to 5s, which would abort the 60s strike
-    stream xread and the unbounded execution pubsub.listen.  Docket overrides
-    socket_timeout to None so blocking reads wait for the server as intended.
-    """
-    async with RedisConnection(redis_url) as connection:
-        assert connection._connection_pool is not None
-        redis_connection = connection._connection_pool.make_connection()
-        assert redis_connection.socket_timeout is None
-
-
-@skip_memory
-@skip_cluster
-async def test_blocking_read_outlasts_redis_py_default_socket_timeout(redis_url: str):
-    """A blocking xread longer than redis-py 8's 5s default socket_timeout waits
-    for the full block instead of raising TimeoutError partway through."""
-    async with RedisConnection(redis_url) as connection:
-        async with connection.client() as r:
-            start = time.monotonic()
-            result = await r.xread({"docket:absent-stream": "$"}, block=6_000)
-            elapsed = time.monotonic() - start
-
-    assert not result
-    assert elapsed >= 5.5
